+++ labs/middleware.py
@@ -4,7 +4,7 @@ from labs.config import settings
 from labs.litellm_service.request import RequestLiteLLM
 from labs.database.embeddings import find_similar_embeddings
 from labs.response_parser.parser import create_file, modify_file, parse_llm_output
-from labs.database.vectorize import vectorize_to_db
+from labs.database.vectorize import vectorize_to_database
 
 logger = logging.getLogger(__name__)
 
@@ -118,7 +118,7 @@ def call_llm_with_context(repo_destination, issue_summary):
         logger.error("issue_summary cannot be empty.")
         raise ValueError("issue_summary cannot be empty.")
 
-    vectorize_to_db(None, repo_destination)
+    vectorize_to_database(None, repo_destination)
     # find_similar_embeddings narrows down codebase to files that matter for the issue at hand.
     context = find_similar_embeddings(issue_summary)
